requirements.txt
openai>=1.0.0
pyyaml>=6.0
sentence-transformers>=2.2.0
torch>=2.0.0
numpy>=1.20.0
pandas>=1.3.0
scikit-learn>=1.0.0
scipy>=1.7.0